Add lignefacture test

+++ src/main/java/com/mycompany/maven_tests/model/lignefacture.java
@@ -1,13 +1,13 @@
 package com.mycompany.maven_tests.model;
 
 public class lignefacture {
-    private Long id;
-    private Long id_facture;
-    private Long id_article;
+    private int id;
+    private int id_facture;
+    private int id_article;
     private int quantity;
     private double sub_total;
 
-    public lignefacture(Long id_facture, Long id_article, int quantity, double sub_total) {
+    public lignefacture(int id_facture, int id_article, int quantity, double sub_total) {
         this.id_facture = id_facture;
         this.id_article = id_article;
         this.quantity = quantity;
@@ -15,19 +15,19 @@ public class lignefacture {
     }
 
     // Getters et Setters
-    public Long getId_facture() {
+    public int getId_facture() {
         return id_facture;
     }
 
-    public void setId_facture(Long id_facture) {
+    public void setId_facture(int id_facture) {
         this.id_facture = id_facture;
     }
 
-    public Long getId_article() {
+    public int getId_article() {
         return id_article;
     }
 
-    public void setId_article(Long id_article) {
+    public void setId_article(int id_article) {
         this.id_article = id_article;
     }
